Refactor wasm into arch ##arch
