+++ libr/anal/meson.build
@@ -112,14 +112,13 @@ r_anal_sources = [
   'p/anal_tricore.c',
   'p/anal_vax.c',
   'arch/vax/vax-dis.c',
-  'p/anal_wasm.c',
+  'p/anal_ws.c',
   # join_paths('arch/whitespace/wsdis.c'),
   'p/anal_x86_cs.c',
   'p/anal_xcore_cs.c',
   'p/anal_z80.c',
   #join_paths('arch','gb','meta_gb_cmt.c'),
   'arch/ebc/ebc_disas.c',
-  'arch/wasm/wasm.c',
   'arch/kvx/kvx-dis.c',
   'arch/kvx/kvx-reg.c',
   'arch/loongarch/gnu/loongarch-coder.c',
