+++ libr/include/r_arch.h
@@ -261,6 +261,7 @@ extern RArchPlugin r_arch_plugin_riscv;
 extern RArchPlugin r_arch_plugin_any_as;
 extern RArchPlugin r_arch_plugin_any_vasm;
 extern RArchPlugin r_arch_plugin_arm;
+extern RArchPlugin r_arch_plugin_wasm;
 extern RArchPlugin r_arch_plugin_x86_nz;
 extern RArchPlugin r_arch_plugin_x86_nasm;
 extern RArchPlugin r_arch_plugin_snes;
