+++ test/db/formats/web_assembly
@@ -85,16 +85,16 @@ CMDS=aa;afl
 EXPECT=<<EOF
 0x00000207    1      3 entry0
 0x00000516    1     19 sym.emscripten_stack_init
-0x0000023d    3    209 sym.main
-0x0000032d    1    126 sym.atoi_via_import
-0x000003b9    1     54 sym.static_int_sum
-0x000003f3    3     14 fcn.000003f3
+0x0000023d    4    209 sym.main
+0x0000032d    2    126 sym.atoi_via_import
+0x000003b9    2     54 sym.static_int_sum
+0x000003f1    3     15 sym._start
 0x000004ab    1      3 sym.__original_main
-0x00000415    6    148 sym.__main_void
-0x000004df    1     12 sym.exit
+0x00000415    9    148 sym.__main_void
+0x000004df    1     11 sym.exit
 0x000004b0    1      1 sym.dummy
-0x000004b5    3     40 sym.libc_exit_fini
-0x000004ed    1      6 sym._Exit
+0x000004b5    5     40 sym.libc_exit_fini
+0x000004ed    1      5 sym._Exit
 0x00000403    1      7 sym.main_2e_1
 0x000004f5    1      3 sym.stackSave
 0x000004fa    1      5 sym.stackRestore
